requirements.txt
python-dotenv
dnspython==1.16.0
pymongo==3.12.0
streamlit
pandas
Pillow
easyocr
torchvision
torch